Create CT4021 - Net bank account calculation
Question 1: Write a program, which calculates the net amount of a bank account based on the transaction log from user input. When the program executes, the user should be able to choose between; Deposit or Withdraw. If user chooses to deposit then he should get another option to select between cash or cheque deposit. Any cash deposit in the account by the user, which is above £1500 should get 0.5% tax deduction. Any cash deposit above £4000 should ask for a reference and payee information. The user should get an allowance of £100 overdraft, however, if there is an overdraft, the account total should present the overdraft amount + 5% bank charges on the overdraft amount.