Create CT4021 - Maths equations
Question 2: Using the maths library,  write a program which solves equations for the user. The program should give user a choice to select amongst different equations the program can solve. A) solve a quadratic equation, B) Calculate area of a circle, C) Calculate circumference of a circle, D) Calculate area of a triangle, E) Volume of a cylinder. Your program should be able to take values from the user and calculate and present the output.